+++ clouds/bigquery/modules/sql/processing/__DELAUNAYGENERIC.sql
@@ -10,7 +10,7 @@ AS ((
         SELECT ST_GEOGPOINT(x, y) AS point FROM (
           SELECT DISTINCT ROUND(ST_X(point), 5) AS x, ROUND(ST_Y(point), 5) AS y
           FROM UNNEST(inputpoints) AS point
-        )
+        ) AS rounded_points
     ),
     points AS (
         SELECT